Ajoute le cas des cantines avec gestionnaire, besoin d'un composant pour gérer l'affichage dynamique du bouton
